update PhyGeoNet (#731)
* Change the directory structure

* update phygeonet code

* update heat equation code

* update heat equation with bc code

* refine phygeonet code

* add callback for seed and logger
